Restructure code and add materials subdir
To make the imports easier from the materials subdir, also restructured
other files. This moves `MPM` to a separate file so as to remove
circular imports for materials module.
