+++ diffmpm/materials/__init__.py
@@ -1,0 +1,3 @@
+from diffmpm.materials._base import _Material
+from diffmpm.materials.simple import SimpleMaterial
+from diffmpm.materials.linear_elastic import LinearElastic
