+++ diffmpm/materials/_base.py
@@ -1,0 +1,48 @@
+import abc
+from typing import Tuple
+
+
+class _Material(abc.ABC):
+    """Base material class."""
+
+    _props: Tuple[str, ...]
+
+    def __init__(self, material_properties):
+        """Initialize material properties.
+
+        Parameters
+        ----------
+        material_properties: dict
+            A key-value map for various material properties.
+        """
+        self.properties = material_properties
+
+    # @abc.abstractmethod
+    def tree_flatten(self):
+        """Flatten this class as PyTree Node."""
+        return (tuple(), self.properties)
+
+    # @abc.abstractmethod
+    @classmethod
+    def tree_unflatten(cls, aux_data, children):
+        """Unflatten this class as PyTree Node."""
+        del children
+        return cls(aux_data)
+
+    @abc.abstractmethod
+    def __repr__(self):
+        """Repr for Material class."""
+        ...
+
+    @abc.abstractmethod
+    def compute_stress(self):
+        """Compute stress for the material."""
+        ...
+
+    def validate_props(self, material_properties):
+        for key in self._props:
+            if key not in material_properties:
+                raise KeyError(
+                    f"'{key}' should be present in `material_properties` "
+                    f"for {self.__class__.__name__} materials."
+                )
